Adjusted output and reworked assigning Usertags

+++ govlearn-api/src/main/java/com/unimuenster/govlearnapi/tags/repository/TagRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface TagRepository extends JpaRepository<Tag, Integer> {
+public interface TagRepository extends JpaRepository<Tag, Long> {
     @Query(value = """
       SELECT t 
       FROM Tag t
